don't accumulate installer
