+++ src/runner/UpdateUtils.cpp
@@ -159,6 +159,10 @@ void ProcessNewVersionInfo(const github_version_info& version_info,
     if (download_update)
     {
         Logger::trace(L"Downloading installer for a new version");
+
+        // Cleanup old updates before downloading the latest
+        updating::cleanup_updates();
+
         if (download_new_version(new_version_info).get())
         {
             state.state = UpdateState::readyToInstall;
